[front/lib/resources] - refactor: move conversation skill management to SkillResource
 - Conversation skill fetching and upserting logic has been moved from ConversationResource to SkillResource
 - Removal of irrelevant imports and codes from ConversationResource as part of clean-up
